use shared superclass for portal access forms
